+++ web-client/src/presenter/sequences/gotoStartCaseSequence.js
