+++ web-client/src/tests/journey/taxpayerNavigatesToCreateCase.js
@@ -1,6 +1,6 @@
 export default test => {
   it('taxpayer sees the procedure types and case types', async () => {
-    await test.runSequence('gotoFilePetitionSequence');
+    await test.runSequence('gotoStartCaseSequence');
     const procedureTypes = test.getState('procedureTypes');
     expect(procedureTypes).not.toBeNull;
     expect(procedureTypes.length).toBeGreaterThan(0);
